Move sorting code out from search_entities method
Implement TrackSearchFilter for TrackSort and move the sorting code
there.

I needed to add collection_uid parameter to apply_to_query method of
TrackSearchFilter. It is a bit odd because we don't use anywhere
the abstraction that the trait provides, yet it forces us to pass
the new parameter even to those implementations that don't use it.
I keep the trait by now, but if we need to complicate it much more
it'll be better to just add a function directly to each type.